requirements.txt
streamlit>=1.32.0
numpy>=1.24.0
numba>=0.58.0
matplotlib>=3.7.0
Pillow>=9.0.0
tensorflow>=2.13.0
